need sometime

+++ src/main/java/mojong/DefaultScoreService.java
@@ -7,7 +7,7 @@ public class DefaultScoreService implements ScoreService {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(DefaultScoreService.class);
 
-	public long shouldPay(long jibendian, Boolean isZhuangJia, Boolean isWinnerZhuangJia, Boolean isFangPao) {
+	public int shouldPay(int jibendian, Boolean isZhuangJia, Boolean isWinnerZhuangJia, Boolean isFangPao) {
 		LOGGER.info("基本点:{}",jibendian);
 		if (isFangPao && isWinnerZhuangJia) {
 			return get100BeiShu(jibendian * 6);
@@ -21,16 +21,16 @@ public class DefaultScoreService implements ScoreService {
 		return get100BeiShu(jibendian * 2);
 	}
 
-	public long getJiBenDian(int fushu, int fanshu) {
+	public int getJiBenDian(int fushu, int fanshu) {
 		int jibendian = fushu * ((int) Math.pow(2, fanshu + 2));
 		return jibendian;
 	}
 
-	public long get100BeiShu(long num) {
+	public int get100BeiShu(int num) {
 		if (num % 100 == 0) {
 			return num;
 		}
-		long num2 = num + 100 - num % 100;
+		int num2 = num + 100 - num % 100;
 		return num2;
 	};
